Improve tests and upgrade Gradle
